feat: :sparkles: add Marvel API key to environment variables
This commit adds the Marvel API key to the environment variables in the app.module.ts file. It uses the ConfigModule from the NestJS Config package to make the key accessible throughout the application.

Refactor: :recycle: update hero.service.ts

This commit updates the hero.service.ts file. It adds a validation check to ensure that the hero being created exists in the Marvel database. If the hero is not found, a BadRequestException is thrown.

feat: :sparkles: add PrismaService to HeroModule providers

This commit adds the PrismaService to the providers array in the hero.module.ts file. This allows the HeroService to use the PrismaService for interacting with the database.

feat: :sparkles: add default value for id field in Hero model

This commit adds a default value for the id field in the Hero model. This ensures that the id field is always populated with a value when creating a new hero.